+++ src/bin/camo.rs
@@ -2,7 +2,7 @@ use std::{net::SocketAddr, str::FromStr};
 
 use clap::Parser;
 
-use camo_rs::{server, Settings};
+use camo_rs::{server, settings::LogFormat, Settings};
 
 async fn shutdown_signal() {
     tokio::signal::ctrl_c()
@@ -12,8 +12,14 @@ async fn shutdown_signal() {
 
 #[tokio::main]
 async fn main() {
-    tracing_subscriber::fmt::init();
     let settings = Settings::parse();
+
+    let subscriber = tracing_subscriber::fmt().with_max_level(settings.log_level.tracing_level());
+    match settings.log_format {
+        LogFormat::Text => subscriber.with_ansi(false).init(),
+        LogFormat::TextColor => subscriber.with_ansi(true).init(),
+        LogFormat::Json => subscriber.json().init(),
+    }
 
     if !(settings.allow_audio || settings.allow_image || settings.allow_video) {
         println!(